requirements.txt
Django>=4.2
djangorestframework
django-crispy-forms
psycopg2-binary
gunicorn
Pillow
numpy
pandas
scikit-learn
flwr
xgboost
joblib
torch>=1.10.0
torchvision
matplotlib
opencv-python
flask
requests
python-dotenv
torchcam
shap
boto3
